Added prop types to components
