+++ src/components/Fixed.tsx
@@ -1,6 +1,8 @@
 import { ICommonProps, SizeUnit, Type } from "../core-types";
 import * as React from "react";
 import { Space } from "./Space";
+import * as PropTypes from "prop-types";
+import { commonProps } from "../core-react";
 
 interface IFixedProps extends ICommonProps {
 	width?: SizeUnit;
@@ -12,3 +14,11 @@ export const Fixed: React.FC<IFixedProps> = ({ width, height, children, ...commo
 		{children}
 	</Space>
 );
+
+Fixed.propTypes = {
+	...commonProps,
+	...{
+		width: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+		height: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+	},
+};
