+++ src/components/Layer.tsx
@@ -1,4 +1,9 @@
 import * as React from "react";
 import { LayerContext } from "../core-react";
+import * as PropTypes from "prop-types";
 
 export const Layer: React.FC<{ zIndex: number }> = (props) => <LayerContext.Provider value={props.zIndex}>{props.children}</LayerContext.Provider>;
+
+Layer.propTypes = {
+	zIndex: PropTypes.number.isRequired,
+};
